% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildingCDM.R
\name{BuildingCDM}
\alias{BuildingCDM}
\title{BuildingCDM Function}
\usage{
BuildingCDM()
}
\description{
This function is the master function which contains Getquery, setNames, connectDB and executeETL functions. 
By executing BuildingCDM function, the user will be asked to enter required information such as the detailed information of DB schemas, then the ETL process will be executed automatically.
}
