% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectDB.R
\name{connectDB}
\alias{connectDB}
\title{connectDB Function}
\usage{
connectDB()
}
\description{
This function allows the user to enter the detailed information for connection with Database. If the connection is successed then "Connection success!" will be showd on Console, otherwise "Connection failed" will be showed.
}
