% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building_NHISNSC_ETL_package.R
\name{executeETL}
\alias{executeETL}
\title{execute ETL function}
\usage{
executeETL(CDM_ddl = TRUE, master_table = TRUE, location = TRUE,
  care_site = TRUE, person = TRUE, death = TRUE,
  observation_period = TRUE, visit_occurrence = TRUE,
  condition_occurrence = TRUE, observation = TRUE,
  drug_exposure = TRUE, procedure_occurrence = TRUE,
  device_exposure = TRUE, measurement = TRUE,
  payer_plan_period = TRUE, cost = TRUE, generateEra = TRUE,
  dose_era = TRUE, indexing = TRUE, constraints = TRUE)
}
\arguments{
\item{CDM_ddl}{}
}
\description{
execute ETL function
}
