% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNames.R
\name{setNames}
\alias{setNames}
\title{setNames Function}
\usage{
setNames()
}
\description{
This function allows to set the names of variables which will be used in executeETL function. The user will be demanded to enter the names of tables and DB schemas.
}
