% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeETL.R
\name{executeETL}
\alias{executeETL}
\title{executeETL Function}
\usage{
executeETL(CDM_ddl = TRUE, import_voca = TRUE, master_table = TRUE,
  location = TRUE, care_site = TRUE, person = TRUE, death = TRUE,
  observation_period = TRUE, visit_occurrence = TRUE,
  condition_occurrence = TRUE, observation = TRUE,
  drug_exposure = TRUE, procedure_occurrence = TRUE,
  device_exposure = TRUE, measurement = TRUE,
  payer_plan_period = TRUE, cost = TRUE, generateEra = TRUE,
  dose_era = TRUE, cdm_source = TRUE, indexing = TRUE,
  constraints = TRUE)
}
\arguments{
\item{CDM_ddl, }{import_voca, master_table, location, care_site, person, death, observation_period, visit_occurrence, condition_occurrence, observation, drug_exposure, procedure_occurrence, device_exposure, measurement, payer_plan_period, cost, generateEra, dose_era, CDM_source, indexing, constraints}
}
\description{
This function allows to execute ETL process.
}
